+++ docs/changelog.rst
@@ -18,6 +18,17 @@ To avail of fixes in an unreleased version, please download a ZIP file
 `directly from GitHub <https://github.com/mitogen-hq/mitogen/>`_.
 
 
+v0.3.42 (2026-02-20)
+--------------------
+
+* :gh:issue:`1451` :mod:`mitogen`: Refactor module whitelist & blacklist with
+  module overrides and blocks. Improve error messages for denied modules.
+* :gh:issue:`1394` :mod:`ansible_mitogen`, :mod:`mitogen`: Add Incus connection
+  support, allowing users who migrated from LXD to Incus to use ansible-mitogen
+* :gh:issue:`1456` :mod:`ansible_mitogen`: Fix Ansible collections
+  :exc:`!mitogen.core.ModuleDeniedByOverridesError`
+
+
 v0.3.41 (2026-02-10)
 --------------------